+++ Module.php
@@ -2,7 +2,15 @@
 
 namespace humhub\modules\socialshare;
 
+use yii\helpers\Url;
+use humhub\modules\socialshare\drivers\BaseDriver;
+
 class Module extends \humhub\components\Module
 {
     public $resourcesPath = 'resources';
+
+    public function getConfigUrl()
+    {
+        return Url::to(['/socialshare/admin']);
+    }
 }