+++ assets/Assets.php
@@ -2,14 +2,30 @@
 
 namespace humhub\modules\socialshare\assets;
 
-use Yii;
 use yii\web\AssetBundle;
 
 class Assets extends AssetBundle
 {
+    /**
+     * @inheritdoc
+     */
     public $sourcePath = '@socialshare/resources';
 
+    /**
+     * @inheritdoc
+     */
+    public $css = [
+        'css/socialshare.css',
+    ];
+
+    /**
+     * @inheritdoc
+     */
+    public $js = [
+        'js/humhub.socialshare.ShareLink.js'
+    ];
+
     public $publishOptions = [
-        'forceCopy' => 'false',
+        'forceCopy' => 'false'
     ];
 }